requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
backports.asyncio.runner==1.2.0
certifi==2025.8.3
exceptiongroup==1.3.0
fastapi==0.116.1
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.6.4
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-asyncio==1.1.0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
redis==6.4.0
sniffio==1.3.1
sqladmin==0.21.0
SQLAlchemy==2.0.43
starlette==0.47.3
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
WTForms==3.1.2
yarl==1.20.1
uvicorn[standard]==0.23.2
